+++ frontend/packages/client/src/components/modals-2/VoteConfirmation.js
