+++ frontend/packages/client/src/components/modals-2/VoteConfirmed.js
